requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.0
asgiref==3.4.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
billiard==3.6.4.0
boto==2.49.0
boto3==1.21.23
botocore==1.24.23
cached-property==1.5.2
celery==5.1.2
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
darksky==0.3
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.12
django-ckeditor==6.2.0
django-crispy-forms==1.13.0
django-js-asset==2.0.0
django-on-heroku==1.1.2
django-rq==2.5.1
django-storages==1.12.3
docutils==0.18.1
frozenlist==1.2.0
gunicorn==20.1.0
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.8.3
jmespath==0.10.0
kombu==5.1.0
mapq==0.3
more-itertools==8.12.0
multidict==5.2.0
packaging==21.3
Pillow==8.4.0
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
pyparsing==3.0.7
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2022.1
redis==4.1.4
requests==2.27.1
rq==1.10.1
s3transfer==0.5.2
sendgrid==6.9.7
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
typing_extensions==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.14.0
yarl==1.7.2
zipp==3.6.0